We now assign item affixes to a item.
As long as the item type is not quest or artifact it can have one of two
affixes. A suffix or a prefix. This is randomly generated on a chance
based on the characters lotting skill after a successful battle.

Fixed a lot of tests that were broken with this approach. Items can
be flushed out with more details.