We now have a adevnture box showing.
When a character starts on a location that is an adventure, that
character will now see an adventure button. Clicking it reveals a new
box that contains the list of adventures for this location as well as
well as two buttons.

One shows the details of the adventure. The other will allow a character
to embark on the adventure.