Started working with livewire.
- Created a form for creating monsters.
  - We can create the monster, edit it's skills and assign a quest item.

Need to continue to test this, as well as write tests for it.

After the tests are done, abstraction need to happen to make the
wizard its own extendable class component.

There is a bug where if you select a skill and then do not click edit
the page wont change when clicking next or previous or anything because
of unseen validation errors.