Started adding enchanting.
This will work the same way as crafting and use the same timer.

This means that one cannot be done at the same time as the other.

Enchanting requires a specific intelligence level as well as
meeting the minimum skill level required.

Aside from that, the same DC check is in place. We use the item enchant cost
to determine cost of the enchantment.