Fight time out is not being utilized.
The thief class is broken and able to kill everything and it's mothers dog.

Kingdom timers are working perfetly, how evr kingdom data when you have buildings in
queue are not working properly. We should see updated data including updated costs.

One way to stop this is only allow them to upgraded the building if its done upgrading.