Update buildings.
When a kingdom game building gets updated we attempt to update
the actual kingdoms buildings for each kingdom that has the building.

There seems to be an issue where the building, when it upgrades, it
is not applying the data properly to the database. Something is not
working and the jobs don't seem to work.