The basics of war are in place.
There are a few bugs to sort out, but we can damage units, buildings
and so on.

Attacking units will also take damage.

We don't currently have any way of doing reports or alerting the user they have
been attacked or are going to be attacked. This aspect is still a work in progress.