Updated tests.
- Admin is now 100% covered.
- Charts is 100% covered.
- Console is 97.32% covered, which is acceptable, because the kernel is untested.

Flare and Game must be increased while other sections such as Exceptions and Providers
are fine as they are.

Added a fixture, for the github test. This fixture uses the new VCR library which
captured the request so that when the tests are run this fixture is run instead of
actually making the call.