We now show a global message for specific quest items that drops.
These quest items are given out by arriving at locations that have them,
fighting monsters or even adventuring and collecting the rewards.

I believe that adventures them selves, not just the monsters, can give quest items
this will have to be updated as well.

Upon recieveing a quest item with an effect on it, a global message will be shown.