Started working on how quests work.
Characters can have boons for a specified amount of time, a lot of this logic is still
missing. This happens when a character uses an item that they crafted by alchemy.

Some skills can be locked, certian quests can unlock these skills.

Quest handleing is almost in place. We just ned to reward XP and handle that
as well as handle what happens when a skill gets unlocked.