Started working on building cost.
You cannot buy building upgrades if the building needs to be rebuilt
or if the building is max level.

Each building has it's own cost. Extra population can be baought for the
building if the population needed is greator then the current kingdom
pop.

The job in question can repeat.