Upgraded to Laravel 9.
Next up is to test php 8.1, since the deployer is now upgraded.

Not sure if the dployer will actually work to deploy, but we can ssh in
now with the new .phar.

Forked my own version of the mjml library that had some minor upgrades.

- May need to think about this in the future, possibly replacing it.

Fixes to the release page. Slight changes to auth login, nothing majour.

Upgraded trust proxies because laravel breaks shit.